+++ apex_legends_api/__init__.py
@@ -1,0 +1,8 @@
+from .ApexLegendsAPI import ApexLegendsAPI
+from .base import Platform, Action
+
+__all__ = [
+    'ApexLegendsAPI',
+    'Platform',
+    'Action'
+]